requirements.txt
python-dotenv==0.21.0
pyTelegramBotAPI==4.12.0
loguru==0.7.0
peewee==3.16.2
beautifulsoup4==4.12.2
requests==2.31.0
webdavclient3==3.14.6
fake-useragent==1.2.1
